+++ CHANGELOG.md
@@ -11,7 +11,11 @@ All notable changes to this project will be documented in this file. The format 
 ### Backend
 
 - Fixed:
+
   - optional fields not being marked as optional in frontend client (missing '?' indicator) ([#1470](https://github.com/bloom-housing/bloom/pull/1470))
+
+- Changed:
+  - User module has been removed and incorporated into Auth module
 
 ### Frontend